[ios]hybird 解压成功

+++ apps/app-template/ios/arsenal/libraries/library-base/src/bundle/util/CXHybirdUtil.swift
@@ -5,15 +5,23 @@ import RxSwift
 class CXHybirdUtil {
 
     static func getRootDir(_ moduleName: String) -> File {
-        return File(CXHybird.localRootDir!, moduleName)
+        let root = File(CXHybird.localRootDir, moduleName)
+        CXLogUtil.v("rootDir=\(root)")
+        return root
     }
 
     static func getZipFile(_ config: CXHybirdModuleConfigModel) -> File {
-        return File(getRootDir(config.moduleName), "\(config.moduleName)-\(config.moduleVersion)\(CXHybird.bundleSuffix)")
+        let parent = getRootDir(config.moduleName)
+        let name = "\(config.moduleName)-\(config.moduleVersion)\(CXHybird.bundleSuffix)"
+        let zipFile = File(parent, name)
+        CXLogUtil.v("zipFile=\(zipFile)")
+        return zipFile
     }
 
-    static func getUnzipDir(_ config: CXHybirdModuleConfigModel) -> File? {
-        return File(getRootDir(config.moduleName), config.moduleVersion)
+    static func getUnzipDir(_ config: CXHybirdModuleConfigModel) -> File {
+        let unzipDir = File(getRootDir(config.moduleName), config.moduleVersion)
+        CXLogUtil.v("unzipDir=\(unzipDir)")
+        return unzipDir.makeDirs()
     }
 
     /**
@@ -44,14 +52,14 @@ class CXHybirdUtil {
         return success
     }
 
-    static internal func unzipToLocal(_ zipFile: File, _ unZipDir: File?) -> Bool {
+    static internal func unzipToLocal(_ zipFile: File?, _ unZipDir: File?) -> Bool {
         var success = false
         CXFileUtil.deleteDirectory(unZipDir?.path)
-//        do {
-        try? CXZipUtil.unzip(zipFile.path ?? "", unZipDir?.path ?? "")
-        success = true
-//        } catch (error:Error) {
-//        }
+        do {
+            try CXZipUtil.unzip(zipFile?.path, unZipDir?.path)
+            success = true
+        } catch {
+        }
         return success
     }
 
@@ -140,19 +148,25 @@ class CXHybirdUtil {
     }
 
     static func copyModuleZipFromAssets(_ moduleName: String, _ primaryConfig: CXHybirdModuleConfigModel?) -> Bool {
+        CXLogUtil.d("--------[copyModuleZipFromAssets]:  start")
         var success = false
-        /*if (primaryConfig != nil) {
+        if (primaryConfig != nil) {
             let zipFile = getZipFile(primaryConfig!)
             let unzipDir = getUnzipDir(primaryConfig!)
             do {
-                try CXFileUtil.deleteFile(zipFile)
-                try CXFileUtil.deleteDirectory(unzipDir)
-                //todo try CXFileUtil.copy(CXBaseApplication.INSTANCE.assets.open("\(CXHybird.assetsDirName)/\(moduleName)-\(primaryConfig!.moduleVersion)\(CXHybird.bundleSuffix)"), zipFile)
+                if (zipFile != nil && zipFile.exists()) {
+                    try CXFileUtil.deleteFile(zipFile)
+                }
+                if (unzipDir.exists()) {
+                    try CXFileUtil.deleteDirectory(unzipDir)
+                }
+                try CXFileUtil.copy(Bundle.main.path(forResource: "\(moduleName)-\(primaryConfig!.moduleVersion)", ofType: "zip")!, zipFile)
                 success = true
-            } catch (exception: Error) {
-                CXLogUtil.e(CXHybird.TAG, "--------[copyModuleZipFromAssets]: 文件不存在", exception)
+            } catch {
+                CXLogUtil.e(CXHybird.TAG, "--------[copyModuleZipFromAssets]: 文件不存在", error)
             }
-        }*/
+        }
+        CXLogUtil.d("--------[copyModuleZipFromAssets]:  end success:\(success)")
         return success
     }
 
@@ -165,45 +179,44 @@ class CXHybirdUtil {
         CXLogUtil.v(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 开始], 当前线程:\(Thread.currentThread().name), 当前时间:\(CXTimeUtil.yMdHmsS(Date(start)))")
         var allConfigList: MutableList<CXHybirdModuleConfigModel> = MutableList<CXHybirdModuleConfigModel>()
         Observable<Any>.create { observer in
-                    CXLogUtil.e(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip] start read")
                     do {
                         let path = Bundle.main.path(forResource: "all", ofType: "json")
-                        print("path", path)
                         let allConfigJsonString = CXFileUtil.readTextFromFile(path)
-                        CXLogUtil.j(CXLogUtil.VERBOSE, allConfigJsonString)
-                        //allConfigList = CXJsonUtil.parse([CXHybirdModuleConfigModel].self, allConfigJsonString) ?? MutableList<CXHybirdModuleConfigModel>()
+//                        CXLogUtil.j(CXLogUtil.VERBOSE, allConfigJsonString)
+
                         allConfigList = CXJsonUtil.parseArray(allConfigJsonString) ?? MutableList<CXHybirdModuleConfigModel>()
                     } catch {
                         CXLogUtil.e(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 开始], 文件不存在, 当前线程:\(Thread.currentThread().name), 当前时间:\(CXTimeUtil.yMdHmsS(Date(start)))", error)
                         observer.onError(error)
                     }
-                    CXLogUtil.e(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip] end   read")
-                    CXLogUtil.j(CXLogUtil.DEBUG, allConfigList)
 
+                    CXLogUtil.d("--------[getConfigListFromAssetsWithCopyAndUnzip:  allConfigList.size = \(allConfigList.size)")
 
                     if (allConfigList.isNotEmpty()) {
                         for (index, config) in allConfigList.enumerated().reversed(){
-                            CXLogUtil.w(index)
                             let zipFile = getZipFile(config)
+                            CXLogUtil.d("copyPrimaryZipFromAssetsSuccess zipFile:\(zipFile) exists=\(zipFile.exists())")
                             let unzipDir = getUnzipDir(config)
+                            CXLogUtil.d("copyPrimaryZipFromAssetsSuccess unzipDir:\(unzipDir)")
 
-//                            let copyStart = System.currentTimeMillis()
-//                            let copyPrimaryZipFromAssetsSuccess = copyModuleZipFromAssets(config.moduleName, config)
-//                            let copyTime = System.currentTimeMillis() - copyStart
-//
-//                            if (copyPrimaryZipFromAssetsSuccess) {
-//                                let unzipStart = System.currentTimeMillis()
-//                                let unzipToLocalSuccess = unzipToLocal(zipFile, unzipDir)
-//                                let unzipTime = System.currentTimeMillis() - unzipStart
-//
-//                                CXLogUtil.v(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 从 assets 拷贝 \(config.moduleName).zip 成功, 解压\(unzipToLocalSuccess ? "成功" : "失败"), 拷贝耗时:\(copyTime) ms, 解压耗时:\(unzipTime) ms, 当前线程:\(Thread.currentThread().name), 当前时间:\(CXTimeUtil.yMdHmsS(Date(start)))")
-//                                if (!unzipToLocalSuccess) {
-//                                    allConfigList.remove(at: index)
-//                                    CXLogUtil.v(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 解压\(config.moduleName).zip 到文件夹失败, 从列表中删除 \(config.moduleName), 当前线程:\(Thread.currentThread().name), 当前时间:\(CXTimeUtil.yMdHmsS(Date(start)))")
-//                                }
-//                            } else {
-//                                CXLogUtil.v(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 从 assets 拷贝 \(config.moduleName).zip 失败, 拷贝耗时:\(copyTime) ms, 当前线程:\(Thread.currentThread().name), 当前时间:\(CXTimeUtil.yMdHmsS(Date(start)))")
-//                            }
+                            let copyStart = System.currentTimeMillis()
+                            let copyPrimaryZipFromAssetsSuccess = copyModuleZipFromAssets(config.moduleName, config)
+                            let copyTime = System.currentTimeMillis() - copyStart
+                            CXLogUtil.d("copyPrimaryZipFromAssetsSuccess --> \(copyPrimaryZipFromAssetsSuccess), zipFile.exists=\(zipFile.exists())")
+
+                            if (copyPrimaryZipFromAssetsSuccess) {
+                                let unzipStart = System.currentTimeMillis()
+                                let unzipToLocalSuccess = unzipToLocal(zipFile, unzipDir)
+                                let unzipTime = System.currentTimeMillis() - unzipStart
+
+                                CXLogUtil.v(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 从 assets 拷贝 \(config.moduleName).zip 成功, 解压\(unzipToLocalSuccess ? "成功" : "失败"), 拷贝耗时:\(copyTime) ms, 解压耗时:\(unzipTime) ms, 当前线程:\(Thread.currentThread().name), 当前时间:\(CXTimeUtil.yMdHmsS(Date(start)))")
+                                if (!unzipToLocalSuccess) {
+                                    allConfigList.remove(at: index)
+                                    CXLogUtil.v(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 解压\(config.moduleName).zip 到文件夹失败, 从列表中删除 \(config.moduleName), 当前线程:\(Thread.currentThread().name), 当前时间:\(CXTimeUtil.yMdHmsS(Date(start)))")
+                                }
+                            } else {
+                                CXLogUtil.v(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 从 assets 拷贝 \(config.moduleName).zip 失败, 拷贝耗时:\(copyTime) ms, 当前线程:\(Thread.currentThread().name), 当前时间:\(CXTimeUtil.yMdHmsS(Date(start)))")
+                            }
                         }
                     }
                     return Disposables.create()
@@ -214,6 +227,9 @@ class CXHybirdUtil {
                     CXLogUtil.v(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 返回解压成功的 allConfigList.size=\(allConfigList.size)], 当前线程:\(Thread.currentThread().name), 当前时间:\(CXTimeUtil.yMdHmsS(Date(start)))")
                     CXLogUtil.v(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 结束]-----------------------------------------------------------------------------------")
                     CXLogUtil.v(CXHybird.TAG, "--------[getConfigListFromAssetsWithCopyAndUnzip: 结束], 当前线程:\(Thread.currentThread().name), 当前时间:\(CXTimeUtil.yMdHmsS(Date())), 一共耗时:\(System.currentTimeMillis() - start)ms")
+
+                    CXFileUtil.printDirs(CXHybird.localRootDir)
+
                     callback(allConfigList)
                 })
     }
